refactored ops out into new file
